requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
requests>=2.31.0
streamlit>=1.28.0
